pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssapy-toolkit"
version = "1.0.0"
description = "SSAPy Toolkit: extensions for orbital mechanics, plotting, and data IO"
authors = [{ name = "Travis R. Yeager", email = "yeagerastro@gmail.com" }]
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = ["orbital-mechanics", "astrodynamics", "visualization", "utilities"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Astronomy",
  "Topic :: Scientific/Engineering :: Visualization",
]

# All runtime dependencies installed by default
dependencies = [
  "astropy",
  "astroquery",
  "cmake",
  "corner",
  "emcee",
  "gitpython",
  "h5py",
  "imageio",
  "imageio-ffmpeg",
  "ipykernel",
  "ipympl",
  "ipython",
  "ipytree",
  "ipyvolume",
  "jplephem",
  "llnl-ssapy",
  "lmfit",
  "matplotlib",
  "mpmath",
  "nexusformat",
  "npy-append-array",
  "numpy",
  "opencv-python",
  "pandas",
  "pickle4",
  "pillow",
  "psutil",
  "pybind11",
  "pyerfa",
  "PyPDF2",
  "python-xlib",
  "rebound",
  "regex",
  "requests",
  "scikit-learn",
  "seaborn",
  "selenium",
  "sympy",
  "tqdm",
  "spacetrack",
  "vg",
  "yapf",
]

[project.optional-dependencies]
# Only dev tools are optional; users install them explicitly with: pip install -e .[dev]
dev = [
  "pytest",
  "pytest-mpi",
  "ipykernel",
  "sphinx",
  "sphinx-rtd-theme",
  "ruff",
]

[project.urls]
Homepage = "https://github.com/yeager7/yeager-utils"
Issues = "https://github.com/yeager7/yeager-utils/issues"

[tool.setuptools]
packages = { find = { where = ["."], include = ["ssapy_toolkit*"] } }
